Load OMP only once in MacOS Python addon.
In MacOS we need to check for possible loaded OMP libraries and load one in case none is.
However this is done each time we run a simulation, even if we had done this check before.
This adds an overhead to each simulation run in the same session. This commit do this check only once.